Move QuotableLambdaOpInvocationHandler to internal package
